IV. More Features  20. Computed Properties  DONE

+++ app/Livewire/ArticleList.php
@@ -3,6 +3,7 @@
 namespace App\Livewire;
 
 use App\Models\Article;
+use Livewire\Attributes\Computed;
 use Livewire\Attributes\Title;
 use Livewire\Component;
 use Livewire\WithPagination;
@@ -14,30 +15,37 @@ class ArticleList extends AdminComponent
 
     public $showOnlyPublished = false;
 
-    public function delete(Article $article) {
-        $article->delete();
-    }
-
-    public function showAll() {
-        $this->showOnlyPublished = false;
-        $this->resetPage(pageName: 'articles-page');
-    }
-
-    public function showPublished() {
-        $this->showOnlyPublished = true;
-        $this->resetPage(pageName: 'articles-page');
-    }
-
-    public function render()
-    {
+    #[Computed/*(persist: true)*/]
+    public function articles() {
         $query = Article::query();
 
         if ($this->showOnlyPublished) {
             $query->where('published', 1);
         }
 
-        return view('livewire.article-list' , [
-            'articles' => $query->paginate(15, pageName: 'articles-page'),
-        ]);
+        return $query->paginate(15, pageName: 'articles-paginator');
     }
+
+
+    public function delete(Article $article) {
+        if ($this->articles->count() < 15 ) {
+            throw new \Exception('You cannot delete the last article');
+        }
+        $article->delete();
+
+        unset($this->articles);
+        cache()->forget('published-count');
+        $this->dispatch('published-count:reset-count'); // added, because it will not work otherwise
+    }
+
+    public function showAll() {
+        $this->showOnlyPublished = false;
+        $this->resetPage(pageName: 'articles-paginator');
+    }
+
+    public function showPublished() {
+        $this->showOnlyPublished = true;
+        $this->resetPage(pageName: 'articles-paginator');
+    }
+
 }
